Modified the Crafting System to allow for Crafting Recipe Scriptable Objects to be looted from various sources (monster drops, quest rewards and overworld lootables). These recipes immediately go to your Crafting Recipe List and persist when saving/changing scenes.
Removed the Animator Override Controllers for weapons. Might revisit later, but this was done so the Player's Animator Override changes when in combat and out of combat instead. New code was added to the PlayerManager to accomplish this.

Started the process of removing the InventoryItem inheritance from our Abilities. Before, Abilities inherited from ActionItem which itself inherited from InventoryItem. Abilities would go to the Inventory when looted or purchased and then acted as any ActionItem would, like a Healing Potion. The removal of the inheritance to ActionItem (and therefore InventoryItem) intends to differentiate Abilities and give Abilities their own UI. When acquired, new Abilities should be able to be "copied" from the UI to the ActionBar via a dragging action. The idea is that the Abilties would persist (or "live") in the Ability UI menu and not the Inventory UI, which it currently does. The Ability should also not be removed from the Ability UI when copied to the ActionBar, but the Abilities should be removable/replaceable on the ActionBar.

Started the process of creating a separate "Ability Shop" akin to a class trainer from popular MMORPGs like World of Warcraft. To start this process, I essentially duplicated the ItemShop configuration and allowed my newly created AbilityItemn to populate the shop. As a barebones approach, this is now already implemented. Modifications to the UI are to come to differentiate the ItemShop from the AbilityShop.

Miscellaneous changes sprinkled in.